requirements.txt
ibm-watson
ibm-cloud-sdk-core
ibm-watsonx-ai
requests
python-dotenv